Refactoring. Dividing CBattleCallback into smaller files with classes.
