+++ lib/spells/CSpellHandler.cpp
@@ -24,7 +24,7 @@
 
 #include "../CStack.h"
 #include "../BattleInfo.h"
-#include "../CBattleCallback.h"
+#include "../battle/CBattleInfoCallback.h"
 #include "../CGameState.h" //todo: remove
 
 #include "../NetPacks.h" //todo: remove
